Feelings

+++ AAUnlimited/External/ExternalClasses/CharacterNpcAiData.h
@@ -22,7 +22,7 @@ public:
 		IllusionArray<float> ai04_01[3];
 		IllusionArray<float> ai04_02[3];
 	}* m_passiveAiData; //this is a pointer somewhere globally, yet every character has its own copy
-	IllusionArray<float> ai01_03[2];
+	IllusionArray<BYTE> ai01_03[2];
 	IllusionArray<float> ai05_00[13];
 	IllusionArray<int> ai05_01[13];	//these are adjusted in some way using the 05_08 data
 	IllusionArray<float> ai05_02[96];
